onedirection/pyproject.toml
[project]
name = "onedirection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.1.0",
    "pydantic>=2.11.7",
]
